Fix webpack warning by making require more explicit
When using requestnetwork.js on a project bundled with webpack, some
warnings of invalid parsing are displayed. This happens because when
webpack finds dynamic require calls like `const x = require('./' + var)`
it will try to load everything on the indicated path, what makes it try
to parse `.md` files and other unrelated stuff.

A good way to prevent that is to tell webpack what type of file you're
expecting when requiring it. This commit adds the `.json` extension on
the `requestnetworkartifacts` module to accomplish this.

Signed-off-by: Vinicius Andrade <vini.andrade.dev@gmail.com>