Extend CircleCI integration (#64)
* Extend CircleCI integration

Extend the CircleCI pipeline in separate jobs to bootstrap the
monorepo, run the library and contract tests and finally generate the
documentation if the tests succeeded.

* Add contract linting and documentation deployment

Use solium to lint the solidium contracts. Deploy the library and
smart contracts documentation to an S3 bucket when all previous steps
were successful.