add prettier with husky+lint-staged to request-logic (#140)
Runs prettier on staged files: Husky runs lint-staged when committing. lint-staged executes tslint --fix and prettier
I decided to go for config files instead of sections in package.json because it will be easier to automate between packages
I let the root editorconfig at 4 spaces to not cause a huge changes on the v1 packages. But the v2 packages will be on 2 spaces