chore(utils): replace default export with named exports (#1023)
* Replace default export with named exports in utils package to enable tree shaking (fixes #1015)

This change replaces the default export in the utils package with named exports, which allows for better tree shaking and smaller build sizes in user code applications. It also enforces the no-default-export ESLint rule to ensure consistent code style

* running refactored files through prehook (#1015)

* Apply changes to utils package to monorepo

This commit applies the changes made to the utils package to the rest of the monorepo. The changes include replacing the default export with named exports and enforcing the no-default-export ESLint rule.

NOTE: Not all changes could be tested (#1029, #1030, #1031)

* Apply changes to utils package to payment-processor/test/payment/any-to-near.test.ts

* feat: tombchain (#1024)

* feat: tombchain (#1024)

* running refactored files through prehook (#1015)

* Apply changes to utils package to monorepo

This commit applies the changes made to the utils package to the rest of the monorepo. The changes include replacing the default export with named exports and enforcing the no-default-export ESLint rule.

NOTE: Not all changes could be tested (#1029, #1030, #1031)

* Apply changes to utils package to payment-processor/test/payment/any-to-near.test.ts

* fix circular dependency by importing from the corresponding package rather from the same package barrel file

* Resolves #1023: Rename ambiguous functions for clarity

As discussed with @alexandre-abrioux and @benjlevesque, many of the functions in the utils module have been refactored and need to be renamed for improved code readability.

Signed-off-by: marcohefti <marco@heftiweb.ch>

* Fix README, identity, and signature file updates
-Removed list of utils from README to avoid staleness
-Changed 'hasError' to 'identityHasError'
-Changed 'sign' back to original name 'sign'
-Changed 'recover' to 'recoverSigner'

Addressed comments by @MantisClone

Signed-off-by: marcohefti <marco@heftiweb.ch>

* Refactor crypto and ec-utils modules

-Removed the named const from crypto-wrapper.ts and exported functions individually
-Removed the named const from ec-utils.ts and exported functions individually
-Renamed recover() to recoverSigner() in ec-utils.ts
-Prefixed functions with 'ec' to prevent duplicate variables
-Reverted normalizeData() to normalize()

Signed-off-by: marcohefti <marco@heftiweb.ch>

Signed-off-by: marcohefti <marco@heftiweb.ch>
Co-authored-by: MantisClone <david.huntmateo@request.network>
Co-authored-by: leoslr <50319677+leoslr@users.noreply.github.com>